bring some things up
