+++ elements/fact.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Fact:
+    title: str
+    text: str
