+++ elements/objective.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Objective:
+    label: str
+    text: str
